Cleanup the geometry engine a bit and add missing registers Also touched up a few other areas

+++ src/gpu_2d.h
@@ -33,28 +33,28 @@ class Gpu2D
         Gpu2D(Memory *memory);
         Gpu2D(Gpu3DRenderer *gpu3DRenderer, Memory *memory);
 
-        void drawScanline(unsigned int line);
+        void drawScanline(int line);
 
         bool is3DEnabled() { return dispCnt & BIT(3); }
 
         uint32_t *getFramebuffer() { return framebuffer; }
 
-        void setExtPalette(unsigned int slot, uint8_t *data) { extPalettes[slot] = data; }
+        void setExtPalette(int slot, uint8_t *data) { extPalettes[slot] = data; }
 
-        uint32_t readDispCnt()              { return dispCnt;      }
-        uint16_t readBgCnt(unsigned int bg) { return bgCnt[bg];    }
-        uint16_t readMasterBright()         { return masterBright; }
+        uint32_t readDispCnt()      { return dispCnt;      }
+        uint16_t readBgCnt(int bg)  { return bgCnt[bg];    }
+        uint16_t readMasterBright() { return masterBright; }
 
         void writeDispCnt(uint32_t mask, uint32_t value);
-        void writeBgCnt(unsigned int bg, uint16_t mask, uint16_t value);
-        void writeBgHOfs(unsigned int bg, uint16_t mask, uint16_t value);
-        void writeBgVOfs(unsigned int bg, uint16_t mask, uint16_t value);
-        void writeBgX(unsigned int bg, uint32_t mask, uint32_t value);
-        void writeBgY(unsigned int bg, uint32_t mask, uint32_t value);
-        void writeBgPA(unsigned int bg, uint16_t mask, uint16_t value);
-        void writeBgPB(unsigned int bg, uint16_t mask, uint16_t value);
-        void writeBgPC(unsigned int bg, uint16_t mask, uint16_t value);
-        void writeBgPD(unsigned int bg, uint16_t mask, uint16_t value);
+        void writeBgCnt(int bg, uint16_t mask, uint16_t value);
+        void writeBgHOfs(int bg, uint16_t mask, uint16_t value);
+        void writeBgVOfs(int bg, uint16_t mask, uint16_t value);
+        void writeBgX(int bg, uint32_t mask, uint32_t value);
+        void writeBgY(int bg, uint32_t mask, uint32_t value);
+        void writeBgPA(int bg, uint16_t mask, uint16_t value);
+        void writeBgPB(int bg, uint16_t mask, uint16_t value);
+        void writeBgPC(int bg, uint16_t mask, uint16_t value);
+        void writeBgPD(int bg, uint16_t mask, uint16_t value);
         void writeMasterBright(uint16_t mask, uint16_t value);
 
     private:
@@ -80,12 +80,12 @@ class Gpu2D
         Gpu3DRenderer *gpu3DRenderer;
         Memory *memory;
 
-        uint32_t rgb5ToRgb6(uint32_t color);
+        uint32_t rgb5ToRgb6(uint16_t color);
 
-        void drawText(unsigned int bg, unsigned int line);
-        void drawAffine(unsigned int bg, unsigned int line);
-        void drawExtended(unsigned int bg, unsigned int line);
-        void drawObjects(unsigned int line);
+        void drawText(int bg, int line);
+        void drawAffine(int bg, int line);
+        void drawExtended(int bg, int line);
+        void drawObjects(int line);
 };
 
 #endif // GPU_2D_H
